Finish color segmentation

+++ color.h
@@ -16,9 +16,9 @@ typedef enum {
 } Color;
 
 void vc_color_print(Color color);
-int vc_rgb_to_hsv_by_color(IVC *src, IVC *dst, int hmin, int hmax, int smin, int smax, int vmin, int vmax);
-int vc_rgb_to_hsv_blue(IVC *src, IVC *dst);
-int vc_rgb_to_hsv_red(IVC *src, IVC *dst);
-int vc_rgb_to_hsv_yellow(IVC *src, IVC *dst);
+long vc_rgb_to_hsv_by_color(IVC *src, IVC *dst, int hmin, int hmax, int smin, int smax, int vmin, int vmax);
+long vc_rgb_to_hsv_blue(IVC *src, IVC *dst);
+long vc_rgb_to_hsv_red(IVC *src, IVC *dst);
+long vc_rgb_to_hsv_yellow(IVC *src, IVC *dst);
 
 #endif /* COLOR_H */
